Move epoch tests to kani
