+++ src/duration/kani_verif.rs
