+++ src/epoch/system_time.rs
@@ -1,3 +1,13 @@
+/*
+ * Hifitime, part of the Nyx Space tools
+ * Copyright (C) 2023 Christopher Rabotin <christopher.rabotin@gmail.com> et al. (cf. https://github.com/nyx-space/hifitime/graphs/contributors)
+ * This Source Code Form is subject to the terms of the Apache
+ * v. 2.0. If a copy of the Apache License was not distributed with this
+ * file, You can obtain one at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Documentation: https://nyxspace.com/
+ */
+
 use crate::{Duration, Epoch, Errors};
 
 /// Converts the webtime Duration into a hifitime Duration.
